Rename API path /api/courses → /api/resources
- git mv packages/middleware/src/routes/api/courses → resources/
  (file names inside the dir keep their old verbs — getCourse.ts /
  upsertCourse.ts / etc. — that's a code-internal name, polish in 7.3b)
- routes.ts registers the same handler tree at /resources prefix
- coursesApi endpoint string flips to "resources"; the JS export name
  stays coursesApi for now so its ~25 callers don't have to move yet
- incrementCourseProgress URL updated to /api/resources/.../incrementProgress

Schema, types, JS variable names (coursesApi, fetchCourses,
SingleCourse component etc.), and component file names stay on the
'course' name — those flip together in 7.3b.
