+++ packages/middleware/src/routes/api/resources/deleteCourse.ts
