+++ packages/middleware/src/routes/api/resources/duplicateCourse.ts
