+++ packages/middleware/src/routes/api/resources/getCourse.ts
