+++ packages/middleware/src/routes/api/resources/incrementCourseProgress.ts
