+++ packages/middleware/src/routes/api/resources/root.ts
