+++ packages/middleware/src/routes/api/resources/routes.ts
