+++ packages/middleware/src/routes/api/routes.ts
@@ -4,7 +4,7 @@ import { FastifyInstance } from "fastify";
 import apiSeed from "./seed";
 import apiClear from "./clearData";
 import apiFormSubmit from "./submitOnboardData";
-import courses from "./courses/routes";
+import courses from "./resources/routes";
 import topics from "./topics/routes";
 import providers from "./providers/routes";
 import domains from "./domains/routes";
@@ -25,7 +25,7 @@ export default async function (server: FastifyInstance) {
   fastify.register(apiClear);
   fastify.register(apiFormSubmit);
   fastify.register(courses, {
-    prefix: "/courses",
+    prefix: "/resources",
   });
   fastify.register(topics, {
     prefix: "/topics",
